refactor: added auth settings to CQRS review commands and queries.Refactored Controller

+++ src/TrustPay.Application/Reviews/Commands/DeleteReview/DeleteReviewCommand.cs
@@ -3,6 +3,7 @@
 using FluentValidation;
 using MediatR;
 using TrustPay.Application.Common.Interfaces;
+using TrustPay.Application.Common.Interfaces.Auth;
 using TrustPay.Application.Common.Interfaces.EntitiesRepo;
 using TrustPay.Domain.Common;
 
@@ -13,22 +14,31 @@ public class DeleteReviewCommandHandler : IRequestHandler<DeleteReviewCommand, R
 {
     private readonly IReviewRepository _reviewRepository;
     private readonly IUnitOfWork _unitOfWork;
+    private readonly ICurrentUserService _currentUserService;
 
-    public DeleteReviewCommandHandler(IReviewRepository reviewRepository, IUnitOfWork unitOfWork)
+    public DeleteReviewCommandHandler(IReviewRepository reviewRepository, IUnitOfWork unitOfWork,ICurrentUserService currentUserService)
     {
         _reviewRepository = reviewRepository;
         _unitOfWork = unitOfWork;
+        _currentUserService = currentUserService;
     }
 
     public async Task<Result> Handle(DeleteReviewCommand request, CancellationToken cancellationToken)
     {
+        var currentUser = _currentUserService.UserId;
+        var isAdmin = _currentUserService.IsAdmin;
         var review = await _reviewRepository.GetByIdAsync(request.Id, cancellationToken);
         if (review is null)
         {
             return Error.NotFound("Review.NotFound", "Отзыв не найден.");
         }
+        if (currentUser != review.AuthorId || !isAdmin)
+        {
+            return Error.Forbidden("Review.Forbidden","Недостаточно прав чтобы удалить отзыв.");
+        }
 
-        _reviewRepository.Delete(review);
+
+            _reviewRepository.Delete(review);
         await _unitOfWork.SaveChangesAsync(cancellationToken);
 
         return Result.Success();
